# cmd/docs/swagger.yaml
definitions:
  model.ContactLine:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  request.CreateContactLine:
    properties:
      description:
        minLength: 1
        type: string
      name:
        minLength: 1
        type: string
    required:
    - description
    - name
    type: object
...
info:
  contact: {}
paths:
  /contact-lines:
    get:
      consumes:
      - application/json
      description: Get all contact lines with pagination
      parameters:
      - description: Page size
        in: query
        name: size
        type: integer
      - description: Start after
        in: query
        name: startAfter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            items:
              $ref: '#/definitions/model.ContactLine'
            type: array
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Get all contact lines
      tags:
      - Contact Line
    post:
      consumes:
      - application/json
      description: Create a contact line
      parameters:
      - description: Contact Line
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CreateContactLine'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create a contact line
      tags:
      - Contact Line
  /contact-lines/{id}:
    patch:
      consumes:
      - application/json
      description: Update a contact line
      parameters:
      - description: Contact Line ID
        in: path
        name: id
        required: true
        type: string
      - description: Contact Line
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CreateContactLine'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update a contact line
      tags:
      - Contact Line
...
swagger: "2.0"